chore: Remove blockagezer and move prover to core/src.
